requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
mcp @ git+https://github.com/modelcontextprotocol/python-mcp.git
